backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic[email]==2.9.2
